feat(kuma-cp) friendly response in K8s mode (#712)
* feat(kuma-cp) friendly response in K8s mode

When kuma-cp is run in Kubernetes mode, all the changes should
happen through `kubectl`. Send an explanation in the response body,
so clients (i.e. `kumactl`) can show it to the user.

Fixes #165

* feat(kumactl) do not show error code 405 on apply/delete

When kuma-cp sends back 405 on resource Create/Update/Delete,
just show the message in the body. This makes the message more
user friendly.

Signed-off-by: Nikolay Nikolaev <nikolay.nikolaev@konghq.com>