feat(test) E2E testing framework (#784)
The initial implementation of a Go-based integration framework. The proposed implementation is based on Terratest and uses Kind to spawn the needed Kubernetes clusters.

The main target is Kubernetes, with the possibility to add Universal mode clusters based in Docker or VMs.

Adds make test/e2e which spawns a set of K8s clusters and invokes the dedicates tests in the test/e2e folder.

This also fixes some Kind testing infra on the way.