Fix e2e Local/Global tests (#826)
Changes the way Local/Global config is handled in kumactl and kuma-cp. Also, fix testing to accommodate the Local/Global specifics.

Introduces SetupDiagnosticsServer as a way to spin up the server that handles readiness and health check probes for K8s. This is independent of setting up the XDS server, which is not needed in Global mode.


* chore(xds) SetupDiagnosticsServer separated in xds

* chore(kumactl) default to kuma-system-global namespace for global mode

* chore(kumactl) disable parts of app.yaml when deploying Global CP

* chore(kuma-cp) reorder the Mode isolating servers and Validations

* chore(kind) expand deployment options with global/local

* chore(kuma-cp) set cfg.Mode before loading the config file

* chore(*) tests for Global/Local port forwarding and GUI check


Signed-off-by: Nikolay Nikolaev <nikolay.nikolaev@konghq.com>