fix(data-planes) access log missing new line
With the TrafficLog policy, using a custom format with the `type: file` makes the
data-planes concatenate all the Access Logs until a new Log is
generated. This bug is not observed when omiting the format.
This is fixed by addind a new line after each Access Log.
For more informations see #820