feat(kumactl) add install loki for log aggregation (#820)
* feat(kumactl) add install loki for log aggregation

Add the Loki stack to `kumactl install` allowing to see logs
directly in Grafana. The yaml for loki was inspired by:
`https://grafana.github.io/loki/charts`

* Add a new install command to kubectl:
  `kubectl install logging`

* Upgrate Grafana version to 7.0.5 that accepts loki as a
  data source.

* tests(kumactl) tests for kumactl install logging

* fix(data-planes) access log missing new line

With the TrafficLog policy, using a custom format with the `type: file` makes the
data-planes concatenate all the Access Logs until a new Log is
generated. This bug is not observed when omiting the format.
This is fixed by addind a new line after each Access Log.
For more informations see #820

* feat(kumactl): bump grafana version to 7.0.5