feat(kuma-cp) http health checks
Introduced possibility to provide HTTP configuration for health checks,
which will send health checks as HTTP requests.

As internal traffic is mostly using HTTP 2, it's the default protocol
used when health checks are configured to use HTTP. To change that
behaviour you can use `http.use_http1` parameter and set it to true.

Signed-off-by: Bart Smykla <bsmykla@vmware.com>
Signed-off-by: Bart Smykla <bartek@smykla.com>