fix(dns) allow resolution of service names with "." (#1448)
When specifying a `kuma.io/service` with a "." DNS resolution would not work
because it would only look at the first part of the DNS name.
This would also cause `foo.bar.mesh` and `foo.mesh` to resolve to the same thing
which may give a false sense of isolation of meshes in the DNS
We now only strips the last part of the DNS name if it matches the `.mesh` domain suffix.

Signed-off-by: Charly Molter <charly@koyeb.com>
(cherry picked from commit de170f1b889d0562321bced10097edfcbba147ce)