feat(ci/cd) wait for check before e2e/integration
As `check` is running in parallel with `build/images` jobs,
and according to my experience looking at some of the last CI/CD
runs `build/images` is taking slightly more time than the `check`
itself, so it doesn't make sense if the only expectation from
not `requiring` it before running `e2e`-related jobs was to speed
it up.

What we are gaining by waiting for `check` to finish is much
quicker feedback when we forgot to run `make check` before pushing
new changes. It's not very rare to happen and without this
requirement, even if `check` fails, the whole spectrum of heavy
e2e tests will still run without a need, as we will have to still
run it again with the `make check` results.

Of course we can manually cancel that kind of job, but to achieve
this we would have to be always present at CircleCI console.

It has significant impact when we are beginning to introduce tests
which are run on automatically created infrastructure (i.e. EKS)
as it introduces real and unnecessary costs.

From my observations, the last CI/CD run times were:
- build (1m 20s) + images (3m 5s) = 4m 25s
- check = 2m 39s

Signed-off-by: Bart Smykla <bartek@smykla.com>