fix(kuma-cp) use ipv4 when ipv4 and ipv6 available
When doing demo, Jakub realised something was wrong and Kuma was
not working on Google's GKE with Kubernetes 1.19.9 and `kuma-init`
containers injected to demos' services were failing without
any logs or errors.

Everything was working fine on GKE with Kubernetes 1.18.x though,
so we realized containers in the newer versions have attached
both, IPv4 and IPv6 IP addresses.

After adding `--verbose` flag to `kuma-init` we found the message:
```
ip6tables-restore v1.8.4 (legacy): ip6tables-restore: unable to
initialize table 'nat'

Error occurred at line: 1
```

so it reassured me the problem is related to IPv6, when tried to
manually play with `ip6tables` it became clear to me that
something is wrong with kernel modules related to iptables for ipv6
and then I figured out that GKE is not supporting IPv6 at all,
even if network interfaces are receiving IPv6 addresses.

Because in our code, if after asking for IP address of network
interface, we would receive IPv6 address first, we were assuming
we can use IPv6.

I rolled back to the solution where we are asking for nic
IPv4 addresses and if there is none, but there is an address
assigned, we are assumming it has to be IPv6.

It means everytime when there will be both - IPv4 and IPv6
available, we will be using IPv4 (unfortunately)

Signed-off-by: Bart Smykla <bartek@smykla.com>