fix(kuma-cp) Clear snapshots from cache on disconnect
Previously we were pushing an empty snapshot on the cache when a dp was disconnecting.
This was causing the dp to get an empty snapshot after a restart.
This would cause listeners to disappear for a short time and causing the traffic going through the dp to fail.
We now clear the snapshot from the dp to avoid this behaviour.

Fix #2176

Signed-off-by: Charly Molter <charly@koyeb.com>