fix(kuma-cp) supported versions fix
As we were using caret ranges `^` instead of tilde ranges `~`
to catch the latest versions (1.1.0, 1.2.0) both of them could
catch the 1.2.0 version. Versions are delivered to our gui
as a hashmap, and when gui on the 1.2.0 version was asking our api
for the supported version of envoy it could receive the response
from range `^1.2.0` and then, everything was fine, and sometimes
it could receive a response from the `^1.1.0` range, and then
it was displaying warning.

It was happening because in the gui, we are iterating over
the hashmap with versions, and there is no consistency in
the order of keys (which is normal).

Also added tests to make sure there is always only one constraint
per version, to not allow similar situation to happen again.

Signed-off-by: Bart Smykla <bartek@smykla.com>