chore(*) upgrade the Cobra package (#2196)
Upgrade Cobra to pick up fixes to the manpage generator.

This also required updating golden output and some changes to how we
match CLI error output. In the latter case, I have dropped the test
discrimination between stdout and stderr since this has changed in the
latest Cobra release. In Kuma, we should not bind our tests too tightly
to Cobra behavior, since we we are going to need to pick up new versions
that are slightly different.

Signed-off-by: James Peach <james.peach@konghq.com>