chore(*) generate DeepCopy interfaces
Enable DeepCopy interface generation for Kubernetes resources. This takes
advantage of the fact that if a type already implementes the DeepCopy
interface, deepcopy-gen will use it. This means that we can wrap the
raw protbuf message fields in a type that supports the interface and
then generate everything else.

This updates #2130.

Signed-off-by: James Peach <james.peach@konghq.com>