chore(*) make the authentication interface type oblivious
There are two proxy types that can be authenticated and in the future
there may be more. Rather than adding a new method for authenticating
each proxy type, switch to authenticating a generic model Resource and
allow each authenticator implentation to type switch internally.

This allows the xDS authentication callbacks to be insensitive to the
DP proxy type, except when fetching an existing resource, which needs
to allocate an instance of the concrete type.

Signed-off-by: James Peach <james.peach@konghq.com>