chore(kuma-cp) use resource types for DataplaneInsight tracking
When we track the Dataplane status, we could be using that to generate
a DataplaneInsight or a ZoneIngressInsight. Previously, we used the
presence of a mesh field to distinguish these cases, but we can take
advantage of the fact that the xDS metadata round-trips the resource
protobuf through the node metadata to determine the exact type when
the we start tracking the xDS client stream. This has the additional
property of not trying to create mis-matched insight resources for any
other types of resources, which ultimately results in errors from the
DataplaneInsightManager.

Signed-off-by: James Peach <james.peach@konghq.com>