chore(*) add initial gateway plugin structure
Add a skeleton gateway plugin that can be enabled by a build-time
feature flag. It is always enabled for tests, but disabled by default
for other builds.

Stub out some structure for Kuma gateway profile generators. We expect
to need a listener generator and a route generator, and it probably make
sense to separate them.

Signed-off-by: James Peach <james.peach@konghq.com>