chore(*) separate proxy template resolvers
Breaking the dependency between the simple proxy template resolver and
the default fallback lets us specify the fallback as a default proxy
template resolver. This in turn, makes it possible for plugins to wrap
the default resolver so that this can be extended.

Moving the default ProxyTemplate resources out of pkg/xds/template
breaks a circular dependency between this package and xds/generator.

Signed-off-by: James Peach <james.peach@konghq.com>