chore(ci) clean up Ginkgo test initialization (#2419)
Move Ginkgo suite initialization to a trivial helper that also sets the
Gomega testing.T. This deletes the (deprecated) custom reporter that was
used in some suites. If we want that back we can add it in one spot so
it gets used consistently.

Signed-off-by: James Peach <james.peach@konghq.com>
(cherry picked from commit 9820f003642afa594072f7358ee067d81cd25977)