fix(*) implement TextMarshaler for JSON keys (#2475)
Add TextMarshaler implementations for types that are uses as map keys.
These types are marshaled to JSON for logging, and the map key conversion
in this case requires encoding.TextMarshaler to be implemented. It is
quite confusing for users to see a marshaling error in the Kuma logs,
since it is difficult to distinguish from a real Kuma error.

This fixes #2374.

Signed-off-by: James Peach <james.peach@konghq.com>
(cherry picked from commit 60d7f97ebbff79bc084f3f83fb66ad7c44aa0128)