feat(*) register the gateway resource type (#2474)
When the build flag is enabled, register the Gateway resource in the
kuma-cp and in kumactl. The latter adds a simple hook to let the plugin
modify the kumactl commands. This is likely to be refactored away as
more of the type bindings are generated.

Signed-off-by: James Peach <james.peach@konghq.com>
(cherry picked from commit b8db547cd1dc5596df069eb3ade5dc2addc4db89)