feat(kuma-cp) add partial support for gatewat traffic routes
Add support for binding TrafficRoute resources to Gateways. This is
based on outbound proxy generator code, but there are still a number of
missing features and outstanding TODO items.

It is possible to configure simple HTTP Gateway use cases and forward
traffic to Kuma services as long as the mesh is not configures with mTLS.
A Gateway can have more than one TrafficRoute resource bound, and this
works correctly by merging the routes in match order.

Signed-off-by: James Peach <james.peach@konghq.com>