fix(kuma-cp): Sort services by name when selecting them
The default sort in k8s is by CreationTimestamp.
In the case of a deployment created in 2 different k8s clusters if
the services were not created in the same order the inbounds would be
generated in a different order which was confusing to the user.

We therefore sort services by names which will always be correct accross zones

Fix #2500

Signed-off-by: Charly Molter <charly.molter@konghq.com>