fix(proxy-template): Correct bad merging of duration
We used to merge 10s and 0.5s into 10.5s this is because
proto.Duration is made of 2 fields (seconds and nanos) and the default
merge method would just add these 2 fields separately.

This patch copies the content of proto.Merge but enables us
to override the merge policy for specific types.

Signed-off-by: Charly Molter <charly.molter@konghq.com>