feat(*) automatically set proxy concurrency (#2691)
This commit adds the `--concurrency` flag to kuma-dp, which is propagated
to Envoy's `--concurrency` flag. On Linux, if there is no explicit
concurrency, we pass `--cpuset-threads` so that Envoy will automatically
scale its worker thread if it is in a cpuset.

In Kubernetes clusters, we can add additional automatic controls for the
Envoy concurrency. We know the amount of CPU resource reserved for the
proxy container, so we can tune that automatically. If the operator knows
better, then can set an explicit concurency value with a pod annotation.

This fixes #1721.

Signed-off-by: James Peach <james.peach@konghq.com>
(cherry picked from commit 1f0cf6ef5acd6bfcee52d09e2968845c925a6cee)