fix(ci) improve golden YAML matcher errors
When the Gomega YAML matcher fails to match, it keeps some information
about the error in the match instance. When the golden output matcher
generates a new matcher instance on each ginkgo call, that internal
state is dumped and we get less error context.

The result of this fix is that the golden YAML matcher will now emit the
first mismatched key in the error, just like the underlying gomega YAML
matcher does.

Signed-off-by: James Peach <james.peach@konghq.com>