chore(k8s) Use a single scheme for all of k8s (#2796)
* chore(k8s) Use a single scheme for all of k8s

We were modifying the same scheme in many places when configuring the
controllers. This is unecessary and is likely to introduce bugs.
We now use the same scheme for everything and modify it at initialization

Signed-off-by: Charly Molter <charly.molter@konghq.com>
(cherry picked from commit fdc5f1a95b1dbbeaa976d3dbec5c8e1966a3afdd)