fix(kuma-cp) remove insight update rate limit burst
The token burst that is configured for the insight update rate limit
allows the update rate to burst up to 50 times before stabilizing to the
desired rate. This would typically happen when kuma-cp starts up and all
the dataplanes reconnect, which is precisely the time when you want the
rate limit to apply for operational stability.

Signed-off-by: James Peach <james.peach@konghq.com>