fix(ci) preserve existing PATH in CircleCI tasks (#2914)
We use "$BASH_ENV" to prepend to "$PATH", since "$BASH_ENV" works like
bashrc and is the recommended way to adjust the path across tasks in
CircleCI.

However, every bash execution will source "$BASH_ENV", so we need to
make sure that we are quoting properly. We want to update "$PATH", not
to reset it to a value that we sampled earlier in the CI workflow. This
particularly matters in the Makefile targets, that assume path
modifications with a Makefile actually work.

This updates #2911.

Signed-off-by: James Peach <james.peach@konghq.com>
(cherry picked from commit ef8da60ce4c7e9d018b9c14a36a089da4709abc3)