chore(ci) enable Gateway for e2e tests (#2890)
First, enable Gateway resources in the build used for end-to-end testing.

The main changes needed to make e2e test pass are to make the control
plane resilient to Gateway resources not having a Kubernetes CRD
representation. For this, we allow the Kubernetes store to silently
return empty or not-found results when asked to retrieve objects of an
unregistered type.

We disable syncing of Gateway resources across zones so that Kubernetes
zones never have to try to store them.

Finally, fix the virtual probes test, which was apparently subject to
race conditions parsing kubectl output.

Signed-off-by: James Peach <james.peach@konghq.com>
(cherry picked from commit 74bcaa6c23104a2d5f0b97d35498d19dbbf182ce)