chore(ci) split e2e deployment options
It's hard to understand which deployment options only apply to the Kuma
control plane, and which apply to arbitrarry applications. Split the
options in two, and fix up all the relevant option types so that we have
strong typing all the way down, and it's obvious which options apply to
which tasks.

Signed-off-by: James Peach <james.peach@konghq.com>