chore(kuma-cp) normalize timeout configurer API (#2934)
All the cluster configurers that have a Kuma resource parameter take
a pointer to the resource object, except for the timeout configurer,
which takes a pointer to an embedded protobuf type. Normalize the API
so that the timeout configurer also takes a resource parameter.

In the subsequent refactoring, the cluster timeout is now only set if it
is specified, instead of being set to 0s. When no timeout is specified,
the Envoy default (15s) will not be applied.

This fixes #2932.

Signed-off-by: James Peach <james.peach@konghq.com>