feat*kumactl) show the effective Dataplane address (#2977)
Show the effective Dataplane address in the Dataplanes listing output.
The effective address is the advertised address if it is present, otherwise
the networking address. The address is one of the most important properties
of the Dataplane, so it's useful to make it more accessible here.

Signed-off-by: James Peach <james.peach@konghq.com>
(cherry picked from commit 4461e71dbe6e52f7a44a0d62f1ed79b811033113)