feat(kuma-cp) add GlobalInsights api endpoint (#3018)
* feat(kuma-cp) add GlobalInsights api endpoint

`/global-insights` is an endpoint which will return aggregated
statistics for global resources, like `Zone`, `ZoneIngress` or
`Mesh`. It's purpose is similar to the `MeshInsights`, but for
global resources, and will be used in our GUI to simplify
displaying statistics of these objects.

I've spent to much time thinking about the approach for adding
this endpoint and finally decided to go with the easiest one,
which is just asking for `Zone`s, `ZoneIngress`es and `Mesh`es
every time endpoint is hitted as the `GlobalInsights` is different
than other resources like `MeshInsights` as there always will be
one. The assumption is also, that there won't be a lot of
the resources we'll be asking for every endpoint hit.

Other solutions considered were:
1. Adding `GlobalInsights` as the real resource, like
   `MeshInsights`, but then we would have to create separate k8s'
   CRD, for just the resource which would exist only one at a time.
   Also, as it's the global resource, reconciliation of it would
   be more complex.
2. Storing the internal data in Config resource (which maps to
   json string in k8s' ConfigMap), but I really think it wouldn't
   be right, as I think it would be not obvious and harder to
   maintain in the future. The complexity of reconciliation would
   be the same as in the option above.

I think this solution is very easy, and if anyhow it will become
performance bottleneck, we can just improve it/refactor it, without
breaking backward compatibility.

Signed-off-by: Bart Smykla <bartek@smykla.com>