feat(kuma-cp) add Gateway TLS termination support
Add support for terminating TLS sessions on Gateway listeners. Most of
the eventual TLS configuration is missing, but TLS 1.2 (and greater)
sessions can be terminated with reasonable defaults.

The approach we take in this PR is to add a new generator to build
the filter chain. Since HTTP and HTTPS protocols get different filter
chains, we separate the logic into two protocol-specific builders. HTTPS
listeners get a filter chain for each hostname, but HTTP listeners only
get a single shared filter chain. Unfortunately, this makes the lifetime
of the filter chain builder a little non-obvious.

Up to two certificate secrets are supported. Each must contain the
private key, the server certificate and the CA certificate chain.

Signed-off-by: James Peach <james.peach@konghq.com>