fix(ci) correct runtime builder ordering
Building a core Runtime object has non-obvious dependencies. The ordering
for building the test runtime was broken because we need to ensure that
the resource manager and its dependencies are present before adding the
datasource loader. The previous order resulted in the datasource loader
failing in tests because it had a nil resource manager.

Signed-off-by: James Peach <james.peach@konghq.com>