fix(ci) Fix gateway test to work with ipv6
missing quotes in yaml template was causing bad syntax when using ipv6
This is because we were getting addresses like "[::]:8080" which without
quotes would look like an array syntax in yaml.
there was a bug in the escape library used for shell commands, replace
the library but a function (probably imperfect but this is just for tests).

Signed-off-by: Charly Molter <charly.molter@konghq.com>