fix(kuma-cp) ensure HTTP router filter ordering (#3082)
The HTTP connection manager filter chain must terminate with the router
filter. Each time a filter is added, the configurer needs to be careful
to prepend itself to the chain to ensure that the router remains last.
This makes it hard to reason about the ordering of filters, since it
will be the reverse of the configurer application order, and it makes
it unnecessarily tricky to write new configurers.

So the easiest approach is for the builder to simply ensure that there
is always a router filter appended when it builds the filter chain.
Since it is a filter chain builder, it's reasonable for it to know
something about filter chain semantics. We might need to do something
more complicated if we ever need to set any fields in the router, but
for now it seems reasonable to stick with simplicity.

Signed-off-by: James Peach <james.peach@konghq.com>