fix(kuma-cp) make cluster names contextually unique (#3098)
For Gateway, it is not good enough to create unique cluster names by
combining the service name and tags. The same service name and tags can be
referenced by different listeners and have different connection policies
applied. When this happens, the different cluster policies are lost
because the cluster resources are uniqified based on the resource name.

The fix is to include a hash the contents of the cluster to in its name to
guarantee that each unique cluster hane has a unique configuration. This
makes the cluster names less friendly for humans, but gives us assurances
that the connection policies always generate distinct clusters.

Signed-off-by: James Peach <james.peach@konghq.com>