chore(kumactl) remove install CRDs filter function
When kumactl tries to install CRDs that aren't already present, we don't
need to filter the names of the CRDs that are already installed to. The
map of installation candidates already contains only Kuma CRDs, and if we
attempt to delete an entry that is not present, that's a harmless no-op.

Signed-off-by: James Peach <james.peach@konghq.com>