chore(kumactl) remove install CRDs filter function (#3139)
When kumactl tries to install CRDs that aren't already present, we don't
need to filter the names of the CRDs that are already installed to. The
map of installation candidates already contains only Kuma CRDs, and if we
attempt to delete an entry that is not present, that's a harmless no-op.

Signed-off-by: James Peach <james.peach@konghq.com>
(cherry picked from commit 07a069ed1f39551f35c0d621148899c4547a62dc)