fix(kuma-cp) validate newly generated xDS snapshots (#3195)
The xDS resource validation is built in to the Go packages, as well as the
Envoy xDS client. Running the validation before we swap the xDS snapshot
in the cache prevents Kuma from sending known-bad configuration to Envoy.

Validation is particularly helpful in tests, since we can ensure that
Kuma is generating valid xDS resources even without a running Envoy to
send them to. This check found a case where the Gateway generator would
generate routes without the require path match.

This fixes #2537.

Signed-off-by: James Peach <james.peach@konghq.com>