chore(build): remove hard-coded type generator outputs
Refactor resource-gen so that each time it runs, it generates exactly one
output. This makes it more obvious which parts of the build are generated
because it's specified directly in the makefiles. This was the original
idea around how this tool should work, but there were previously some
issues around generating the protobuf type information that needed
multiple outputs. Now that those are resolved, we can simplify the
generator again.

This updates #3479.

Signed-off-by: James Peach <james.peach@konghq.com>