tests(kuma-cp): add initial Gateway tests for Kubernetes
Start adding Kubernetes e2e tests for the builting Gateway. We take
advantage of the fact that GatewayInstance supports ClusterIP services
to be able to rendevous with the gateway address. The intention is to
duplicate the Universal cluster tests (where appropriate), so there is
probably some scope for refactoring and consolidation in the future.

This updates #3510.

Signed-off-by: James Peach <james.peach@konghq.com>