fix(build): move API makefile to the main build system
Merge `api/Makefile` back to `mk/api.mk`. This consolidates the build and
removes some duplication of protobuf configuration.  The main issue is
that protoc binds the proto search path and the output path in surprising
and undocumented ways. So, depending on the context, we have to either
change into the "api" directory, or add it to the search path.

This fixes #2495.

Signed-off-by: James Peach <james.peach@konghq.com>