fix(kumactl): print meta before spec for Kuma resources
When printing Kuma resources as YAML, it's a better user experience to
show the metadata fields (including the type and name) up front, rather
than the fields being in an arbitrary order.

This change is not a complete fix for the issue, but by tweaking the
kumactl get printer, we can make a small change that improves the user
experience in most cases.

This updates #679.

Signed-off-by: James Peach <james.peach@konghq.com>