chore(e2e): switch Kubernetes update helper to use apply
The Kubernetes cluster object update helper works by doing a "kubectl
get", updating the deserialized result, then doing a "kubectl replace".
The intention is to set exactly what the test specifies without worrying
about patch and merge behavior.

Howwever, replacement requires the resource version to be current and when
Kuma converts Kubernetes objects to internal API resources, it does not
preserve all the Kubernetes metadata. It seems like this can lead to test
failures where the wrong resource version is used, though my understanding
was that kubectl internally ought to fetch the latest resource version
(though maybe it only does that once, and can lose race conditions).

At any rate, switching to "kubectl apply" ought to be OK for the current
uses of this mechanism, and should be more robust in CI.

Signed-off-by: James Peach <james.peach@konghq.com>