chore(e2e): consolidate gateway tests
Refactor the Gateway e2e tests so that they follow a more consistent code
structure and we can hoist out common parts of the tests. This makes it
easier to see what tests are enabled for universal and gateway zones, and
reduces the amount of code we need to copy between the two test suites.

Updates #3510.

Signed-off-by: James Peach <james.peach@konghq.com>