feat(*): change zoneegress token to zone token
After the discussion with a team, we decided that we don't want to
introduce another type of token for zone egresses, and it would be
good to have one token type (ZoneToken), which currently will be
used only for zone egresses, but in the futere iterations
it should replace tokens used to identify dataplanes and ingresses
in zone as well.

As zone token with wide scope could introduce potential risk of
impersonating other resources, we decided that zones will receive
only public key of the zone token signing key, which wil be enough
to validate identities of the resources, but not to sign
the new tokens. It means that you want be able to generate
zone tokens with zone control planes, and you will have to do it
using global control plane only.

Signed-off-by: Bart Smykla <bartek@smykla.com>