feat(kuma-cp): provide better message when running with an in-memory database (#3982)
* feat(kuma-cp): provide better message when running with an in-memory database

When kuma-cp runs with an in-memory database and an instance gets restarted, dataplanes
cannot be authorized because the private key used to sign dataplane token has changed.
Added log to kuma-cp that during start with in-memory database inform that state is
not preserved. Also, kuma-cp returns in XDS response information that when running
in-memory database doesn't keep state between restarts.

Signed-off-by: Łukasz Dziedziak <lukidzi@gmail.com>
Signed-off-by: Sally Blich <sally.blich@walkme.com>