feat(kumactl): add --values and --set to kumactl install control-plane
This makes kumactl a lot closer to what helm does which is the end is what we
use underneath.
We also add move validation to the helm chart so that we have this for
users of the helm chart too.

I believe that in future version we'd simply drop the old arguments but
let's keep it as is until docs and tests are updated to use this new method

Fix #1116

Signed-off-by: Charly Molter <charly.molter@konghq.com>