feat(*): support ARM Linux/Darwin for dev/tools (#4199)
added support for dev/tools installation on arm linux/osx. Separate installation of etcd which works on osx, kube-apiserver which is not always included in kubebuilder package. Shellcheck is installed in a version that works on both ends. protoc and helm also come with version that works on both arch.

Signed-off-by: Łukasz Dziedziak <lukidzi@gmail.com>