feat(helm): mv kds config into tls section
This was tricky, because we had to get rid of the enabled flag in order
to maintain backwards compatibility. Because in the past, we only turned
on the kds cert checking code if a name was set, but now we want to also
turn it on if no name is set but create is true.

Signed-off-by: Michael Younkin <michael@goodrx.com>