fix(kumactl): use the same client in `kumactl apply`
We were creating a new http client for each resource which
was causing an out of fd error when applying a lot of entities at once.

This moves client creation outside of the loop and ensure context is passed in
everywhere.

Also fix --dry-run to output a valid set of yaml with `---` separators

Signed-off-by: Charly Molter <charly.molter@konghq.com>