fix(xds): don't read metadata in ProxyBuilders (#5414)
We were using a `DataplaneMetadataTracker` in all proxy builders.
This was causing issues as the proxy may disconnect while reconciliation is in progress.

We now set the proxy metadata in the DataplaneWatchdog so that it's read once only and nil checked.
We therefore protect ourselves against this race

Signed-off-by: Charly Molter <charly.molter@konghq.com>
(cherry picked from commit 39ba902fa3d60c3c7ec685e3ee8d6125eec70aa5)

# Conflicts:
#	pkg/api-server/inspect_endpoints.go
#	pkg/xds/sync/components.go
#	pkg/xds/sync/egress_proxy_builder.go
#	pkg/xds/sync/ingress_proxy_builder.go