chore(deps): install dev tools and split if more repos (#5528)
Previously code was placed directly in Makefile and thanks to this we didn't
have to pass TOOLS_DEPS_DIRS as a parameter to the script. When in another repository
we are adding to env TOOLS_DEPS_DIRS another folder to scan items the value that enters the scripts is
split by space which bash later recognizes as 2 separate parameters. That caused the other dir to be treated as GOOS.

Signed-off-by: Łukasz Dziedziak <lukidzi@gmail.com>