feat(e2e/universal): display full logs after fail
Logs which were for convinience saved under /tmp/e2e, now are
wrapped by directories which names constists of suite start time.

Each universal component logs are also placed under directory with
the name of the test (which is easier to find when needed).

When test fails, the logs are written to GinkgoWriter, so in CI
we'll have access to them stright away.

Signed-off-by: Bart Smykla <bartek@smykla.com>