feat(kuma-cp): introduce KDS sync from Global to Zone service with delta xDS (#6278)
Introduced new grpc service KDSSyncService which has currently one rpc GlobalToZoneSync which is responsible for synchronizing resources from Global To Zone. The implementation uses go-control-plane cache and incremental xDS server. Instead of sending SOTW in each DiscoveryResponse, the server sends only changed/new resources. ResourceNamesSubscribe has a value of * which allows getting updates on all the resources.

Changes in sync store:

When Zone CP doesn't have entries in deltaInitialState that means it is a first request and the whole state needs to be propagated from Global to Zone. Based on this we are using the same mechanism we are using in the previous KDS implementations which compares upstream and downstream and removed resources that are out of sync. That allows protection from a state where some resources are still in a zone cp but shouldn't be.
When there are entries in DeltaDiscoveryResponse.RemovedResourceNames we are removing only these resources from the store.
When RemovedResourceNames has the name of the resource and in Resources there is also the resource with the same name we need to update it.
Resources are saved in the system namespace
Other changes

When the client returns NACK, the go-control-plane server for SOTW was retrying requests. That is not correct behavior and delta doesn't do this. That is a problem when the control plane runs on K8s and control-planes are not ready before the first requests arrive from webhooks. To avoid this issue I've introduced the retryer which reset the versions of the resources which later forces watches to send the response again. Setting a different version of the snapshot won't work because delta xDS calculate the version of the resource because of the hash of the object. If an object hasn't changed, the update won't be sent.

So far everything is behind a flag and more tests could be run when we implement ZoneToGlobalSync

IMPORTANT CHANGE:

Namespace resources are not stored in the default namespace anymore. By default, we are using a system namespace.

--------------

Signed-off-by: Lukasz Dziedziak <lukidzi@gmail.com>