ci(helm): no duplicates in helm chart package (#6512)
Chart Releaser is always packaging dependencies specified in Chart.yaml
file first. The archives with packaged dependencies lands inside "charts"
directory (i.e. deployments/[...]/charts/[dependency_name]-[version].tgz).
When building the final chart package, it includes inside, the content of
the "charts" directory. When it finds an archive there, it flattens it and
includes in the final package as well. It means if the chart have any
dependencies, the final chart's archive will contain duplicated manifests.
To mitigate this issue we are unarchiving (which would remove duplicates)
and archiving the chart again.

Signed-off-by: Bart Smykla <bartek@smykla.com>